+++ packages/vaex-core/vaex/dataframe.py
@@ -1458,13 +1458,12 @@ class DataFrame(object):
         waslist, [expressions, ] = vaex.utils.listify(expression)
         limits = []
         for expr in expressions:
-            subspace = self(expr)
-            limits_minmax = subspace.minmax()
-            vmin, vmax = limits_minmax[0]
+            limits_minmax = self.minmax(expr)
+            vmin, vmax = limits_minmax
             size = 1024 * 16
-            counts = subspace.histogram(size=size, limits=limits_minmax)
+            counts = self.count(binby=expr, shape=size, limits=limits_minmax)
             cumcounts = np.concatenate([[0], np.cumsum(counts)])
-            cumcounts /= cumcounts.max()
+            cumcounts = cumcounts / cumcounts.max()
             # TODO: this is crude.. see the details!
             f = (1 - percentage / 100.) / 2
             x = np.linspace(vmin, vmax, size + 1)
@@ -2028,7 +2027,10 @@ class DataFrame(object):
             data = column[0:1]
             dtype = data.dtype
         else:
-            data = self.evaluate(expression, 0, 1, filtered=True, internal=True, parallel=False)
+            try:
+                data = self.evaluate(expression, 0, 1, filtered=False, internal=True, parallel=False)
+            except:
+                data = self.evaluate(expression, 0, 1, filtered=True, internal=True, parallel=False)
             dtype = data.dtype
         if not internal:
             if dtype != str_type:
@@ -3692,6 +3694,9 @@ class DataFrame(object):
         if not hidden and virtual and regex is None:
             return [k for k in self.column_names if not k.startswith('__')]  # also a quick path
         return [name for name in self.column_names if column_filter(name)]
+
+    def __bool__(self):
+        return True  # we are always true :) otherwise Python might call __len__, which can be expensive
 
     def __len__(self):
         """Returns the number of rows in the DataFrame (filtering applied)."""